Refactor common code from analyze.dart and test.dart into run_command.dart
