+++ dev/bots/test.dart
@@ -3,10 +3,11 @@
 // found in the LICENSE file.
 
 import 'dart:async';
-import 'dart:convert';
 import 'dart:io';
 
 import 'package:path/path.dart' as path;
+
+import 'run_command.dart';
 
 typedef Future<Null> ShardRunner();
 
@@ -16,17 +17,6 @@ final String dart = path.join(flutterRoot, 'bin', 'cache', 'dart-sdk', 'bin', Pl
 final String pub = path.join(flutterRoot, 'bin', 'cache', 'dart-sdk', 'bin', Platform.isWindows ? 'pub.bat' : 'pub');
 final String pubCache = path.join(flutterRoot, '.pub-cache');
 final List<String> flutterTestArgs = <String>[];
-final bool hasColor = stdout.supportsAnsiEscapes;
-
-final String bold = hasColor ? '\x1B[1m' : '';
-final String red = hasColor ? '\x1B[31m' : '';
-final String green = hasColor ? '\x1B[32m' : '';
-final String yellow = hasColor ? '\x1B[33m' : '';
-final String cyan = hasColor ? '\x1B[36m' : '';
-final String reset = hasColor ? '\x1B[0m' : '';
-final String redLine = '$red━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━$reset';
-const String arrow = '⏩';
-const String clock = '🕐';
 
 const Map<String, ShardRunner> _kShards = <String, ShardRunner>{
   // 'analyze': _analyzeRepo, handled by analyze.dart
@@ -137,7 +127,7 @@ Future<Null> _runSmokeTests() async {
         printOutput: false,
         timeout: _kShortTimeout,
       ),
-      _runCommand(flutter,
+      runCommand(flutter,
         <String>['drive', '--use-existing-app', '-t', path.join('test_driver', 'failure.dart')],
         workingDirectory: path.join(flutterRoot, 'packages', 'flutter_driver'),
         expectNonZeroExit: true,
@@ -233,7 +223,7 @@ Future<Null> _pubRunTest(
         toolsArgs += ' --enable-asserts';
     pubEnvironment['FLUTTER_TOOL_ARGS'] = toolsArgs.trim();
   }
-  return _runCommand(
+  return runCommand(
     pub, args,
     workingDirectory: workingDirectory,
     environment: pubEnvironment,
@@ -250,69 +240,6 @@ class EvalResult {
   final String stdout;
   final String stderr;
   final int exitCode;
-}
-
-String elapsedTime(DateTime start) {
-  return new DateTime.now().difference(start).toString();
-}
-
-Future<Null> _runCommand(String executable, List<String> arguments, {
-  String workingDirectory,
-  Map<String, String> environment,
-  bool expectNonZeroExit = false,
-  int expectedExitCode,
-  String failureMessage,
-  bool printOutput = true,
-  bool skip = false,
-  Duration timeout = _kLongTimeout,
-}) async {
-  final String commandDescription = '${path.relative(executable, from: workingDirectory)} ${arguments.join(' ')}';
-  final String relativeWorkingDir = path.relative(workingDirectory);
-  if (skip) {
-    _printProgress('SKIPPING', relativeWorkingDir, commandDescription);
-    return null;
-  }
-  _printProgress('RUNNING', relativeWorkingDir, commandDescription);
-
-  final DateTime start = new DateTime.now();
-  final Process process = await Process.start(executable, arguments,
-    workingDirectory: workingDirectory,
-    environment: environment,
-  );
-
-  Future<List<List<int>>> savedStdout, savedStderr;
-  if (printOutput) {
-    await Future.wait(<Future<void>>[
-      stdout.addStream(process.stdout),
-      stderr.addStream(process.stderr)
-    ]);
-  } else {
-    savedStdout = process.stdout.toList();
-    savedStderr = process.stderr.toList();
-  }
-
-  final int exitCode = await process.exitCode.timeout(timeout, onTimeout: () {
-    stderr.writeln('Process timed out after $timeout');
-    return expectNonZeroExit ? 0 : 1;
-  });
-  print('$clock ELAPSED TIME: $bold${elapsedTime(start)}$reset for $commandDescription in $relativeWorkingDir: ');
-  if ((exitCode == 0) == expectNonZeroExit || (expectedExitCode != null && exitCode != expectedExitCode)) {
-    if (failureMessage != null) {
-      print(failureMessage);
-    }
-    if (!printOutput) {
-      stdout.writeln(utf8.decode((await savedStdout).expand((List<int> ints) => ints).toList()));
-      stderr.writeln(utf8.decode((await savedStderr).expand((List<int> ints) => ints).toList()));
-    }
-    print(
-      '$redLine\n'
-      '${bold}ERROR:$red Last command exited with $exitCode (expected: ${expectNonZeroExit ? (expectedExitCode ?? 'non-zero') : 'zero'}).$reset\n'
-      '${bold}Command:$cyan $commandDescription$reset\n'
-      '${bold}Relative working directory:$red $relativeWorkingDir$reset\n'
-      '$redLine'
-    );
-    exit(1);
-  }
 }
 
 Future<Null> _runFlutterTest(String workingDirectory, {
@@ -340,17 +267,13 @@ Future<Null> _runFlutterTest(String workingDirectory, {
     }
     args.add(script);
   }
-  return _runCommand(flutter, args,
+  return runCommand(flutter, args,
     workingDirectory: workingDirectory,
     expectNonZeroExit: expectFailure,
     printOutput: printOutput,
     skip: skip,
     timeout: timeout,
   );
-}
-
-void _printProgress(String action, String workingDir, String command) {
-  print('$arrow $action: cd $cyan$workingDir$reset; $yellow$command$reset');
 }
 
 Future<Null> _verifyVersion(String filename) async {
